Split into separate classes, add textarea
